refactor(oracle): migrate wifi eapol and rsn to wifi plugins
