+++ tools/oracle/engine/backends/scapy/normalize.py
@@ -63,8 +63,9 @@ _LAYER_ALIASES: dict[str, str] = {
     "BTLE_ADV": "ble_adv",
     "BTLE_ADV_IND": "ble_adv",
     "Dot11": "dot11",
-    "Dot11EltRSN": "rsn",
-    "EAPOL": "eapol",
+    # The ``Dot11EltRSN`` -> ``rsn`` and ``EAPOL`` -> ``eapol`` decode aliases moved
+    # to the migrated ``protocols/wifi.py`` plugin (their ``ScapyProtocol.layer_aliases``);
+    # ``_normalize_layer_name`` resolves them from ``_registered_layer_aliases``.
     "IGMP": "igmp",
     "IGMPv3": "igmp",
     "IGMPv3gr": "igmp_group_record",
