+++ tools/oracle/engine/backends/wireshark/normalize.py
@@ -16,7 +16,6 @@ from .decode_helpers import (
     _field,
     _field_list,
     _fields_from_aliases,
-    _hex_bytes,
     _layer,
     _layer_any,
     _normalize_root_name,
@@ -331,10 +330,6 @@ def _normalize_protocol_fields(
         return plugin.normalize(layers, source_hex=source_hex)
     if layer_name == "dot11":
         return _normalize_dot11(_layer(layers, "wlan"))
-    if layer_name == "eapol":
-        return _normalize_eapol(_layer(layers, "eapol"))
-    if layer_name == "rsn":
-        return _normalize_rsn(_layer_any(layers, "wlan_mgt.rsn", "wlan.rsn"))
     if layer_name == "ipv6_hop_by_hop":
         return _normalize_ipv6_options_header(
             _layer_any(layers, "ipv6.hopopts", "ipv6_hopopts"),
@@ -474,81 +469,11 @@ def _normalize_dot11(layer: JSONObject) -> JSONObject:
     return output
 
 
-def _normalize_eapol(layer: JSONObject) -> JSONObject:
-    output = _fields_from_aliases(
-        layer,
-        {
-            "version": ("eapol.version",),
-            "packet_type": ("eapol.type",),
-            "body_length": ("eapol.len", "eapol.length"),
-            "descriptor_type": ("eapol.keydes.type",),
-            "key_information": ("eapol.keydes.key_info",),
-            "key_length": ("eapol.keydes.key_len",),
-            "replay_counter": ("eapol.keydes.replay_counter",),
-            "key_nonce": ("eapol.keydes.nonce",),
-            "key_iv": ("eapol.keydes.key_iv",),
-            "key_rsc": ("eapol.keydes.key_rsc",),
-            "key_id": ("eapol.keydes.key_id",),
-            "key_mic": ("eapol.keydes.key_mic",),
-            "key_data_length": ("eapol.keydes.keydes_data_len",),
-            "key_data": ("eapol.keydes.data",),
-        },
-    )
-    _parse_int_fields(
-        output,
-        "version",
-        "packet_type",
-        "body_length",
-        "descriptor_type",
-        "key_information",
-        "key_length",
-        "replay_counter",
-        "key_data_length",
-    )
-    for name in ("key_nonce", "key_iv", "key_rsc", "key_id", "key_mic", "key_data"):
-        value = output.get(name)
-        if isinstance(value, str):
-            output[name] = {"hex": _hex_bytes(value)}
-    return output
-
-
-def _normalize_rsn(layer: JSONObject) -> JSONObject:
-    output = _fields_from_aliases(
-        layer,
-        {
-            "element_id": ("wlan.rsn.tag", "wlan.rsn.element_id"),
-            "length": ("wlan.rsn.length",),
-            "version": ("wlan.rsn.version",),
-            "capabilities": ("wlan.rsn.capabilities",),
-        },
-    )
-    _parse_int_fields(output, "element_id", "length", "version", "capabilities")
-    group = _rsn_suite_from_layer(
-        layer,
-        kind="cipher",
-        value_names=("wlan.rsn.gcs", "wlan.rsn.group_cipher_suite"),
-        oui_names=("wlan.rsn.gcs.oui",),
-        type_names=("wlan.rsn.gcs.type",),
-    )
-    if group is not None:
-        output["group_cipher_suite"] = group
-    pairwise = _rsn_suite_list_from_layer(
-        layer,
-        kind="cipher",
-        value_names=("wlan.rsn.pcs", "wlan.rsn.pcs.type"),
-        oui_names=("wlan.rsn.pcs.oui",),
-    )
-    if pairwise:
-        output["pairwise_cipher_suites"] = pairwise
-    akms = _rsn_suite_list_from_layer(
-        layer,
-        kind="akm",
-        value_names=("wlan.rsn.akms", "wlan.rsn.akms.type"),
-        oui_names=("wlan.rsn.akms.oui",),
-    )
-    if akms:
-        output["akm_suites"] = akms
-    return output
+# The ``eapol`` and ``rsn`` tshark normalizers (``_normalize_eapol`` /
+# ``_normalize_rsn`` and the RSN suite-selector helper cluster) moved to
+# ``protocols/wifi.py`` and are registered in ``WIRESHARK_REGISTRY`` (so
+# ``_normalize_protocol_fields`` routes the ``eapol`` / ``rsn`` layers to the
+# plugin's ``normalize``).
 
 
 # The DHCP tshark normalizer ``_normalize_dhcp`` and its ``_dhcp_flags`` /
@@ -594,147 +519,11 @@ def _copy_string_field(
         output[target] = value
 
 
-def _rsn_suite_from_layer(
-    layer: JSONObject,
-    *,
-    kind: str,
-    value_names: tuple[str, ...],
-    oui_names: tuple[str, ...],
-    type_names: tuple[str, ...],
-) -> JSONObject | None:
-    value = _field(layer, *value_names)
-    suite = _rsn_suite_from_value(value, kind=kind)
-    if suite is not None:
-        return suite
-    oui = _string_field(layer, *oui_names)
-    suite_type = _parse_int(_field(layer, *type_names))
-    if oui is None or suite_type is None:
-        return None
-    return _rsn_suite_from_oui_type(oui, suite_type, kind=kind)
-
-
-def _rsn_suite_list_from_layer(
-    layer: JSONObject,
-    *,
-    kind: str,
-    value_names: tuple[str, ...],
-    oui_names: tuple[str, ...],
-) -> list[JSONObject]:
-    values = _field_list(layer, *value_names)
-    suites = [
-        suite
-        for suite in (_rsn_suite_from_value(value, kind=kind) for value in values)
-        if suite is not None
-    ]
-    if suites:
-        return suites
-    oui_values = [str(value) for value in _field_list(layer, *oui_names)]
-    if not oui_values:
-        oui_values = ["00:0f:ac"] * len(values)
-    output: list[JSONObject] = []
-    for index, value in enumerate(values):
-        suite_type = _parse_int(value)
-        if suite_type is None:
-            continue
-        oui = oui_values[index] if index < len(oui_values) else oui_values[0]
-        suite = _rsn_suite_from_oui_type(oui, suite_type, kind=kind)
-        if suite is not None:
-            output.append(suite)
-    return output
-
-
-def _rsn_suite_from_oui_type(oui: str, suite_type: int, *, kind: str) -> JSONObject | None:
-    try:
-        oui_bytes = bytes.fromhex(_hex_bytes(oui))
-    except ValueError:
-        return None
-    if len(oui_bytes) < 3:
-        return None
-    return _rsn_suite_selector(oui_bytes[:3] + bytes([suite_type & 0xFF]), kind=kind)
-
-
-def _rsn_suite_from_value(value: object, *, kind: str) -> JSONObject | None:
-    if value is None:
-        return None
-    parsed = _parse_int(value)
-    if parsed is not None:
-        return _rsn_suite_selector(b"\x00\x0f\xac" + bytes([parsed & 0xFF]), kind=kind)
-    if not isinstance(value, str):
-        return None
-    raw_hex = _hex_bytes(value)
-    if len(raw_hex) < 8:
-        return None
-    try:
-        raw = bytes.fromhex(raw_hex[:8])
-    except ValueError:
-        return None
-    return _rsn_suite_selector(raw, kind=kind)
-
-
-def _rsn_suite_selector(raw: bytes, *, kind: str) -> JSONObject:
-    selector = bytes(raw)
-    label = _rsn_cipher_label(selector) if kind == "cipher" else _rsn_akm_label(selector)
-    output: JSONObject = {
-        "selector": selector.hex(),
-        "oui": selector[:3].hex(),
-        "suite_type": selector[3],
-    }
-    if label is not None:
-        output["label"] = label
-    return output
-
-
-def _rsn_cipher_label(selector: bytes) -> str | None:
-    if selector[:3] != b"\x00\x0f\xac":
-        return None
-    return {
-        0: "use-group",
-        2: "tkip",
-        4: "ccmp-128",
-        6: "aes-128-cmac",
-        7: "no-group-addressed",
-        8: "gcmp-128",
-        9: "gcmp-256",
-        10: "ccmp-256",
-        11: "bip-gmac-128",
-        12: "bip-gmac-256",
-        13: "bip-cmac-256",
-        18: "ccm-star",
-    }.get(selector[3])
-
-
-def _rsn_akm_label(selector: bytes) -> str | None:
-    if selector[:3] != b"\x00\x0f\xac":
-        return None
-    return {
-        1: "802.1x",
-        2: "psk",
-        3: "ft-802.1x",
-        4: "ft-psk",
-        5: "802.1x-sha256",
-        6: "psk-sha256",
-        7: "tdls",
-        8: "sae",
-        9: "ft-sae",
-        10: "ap-peer-key",
-        11: "802.1x-suite-b",
-        12: "802.1x-suite-b-192",
-        13: "ft-802.1x-sha384-cmp-256",
-        14: "fils-sha256",
-        15: "fils-sha384",
-        16: "ft-fils-sha256",
-        17: "ft-fils-sha384",
-        18: "owe",
-        19: "ft-psk-sha384",
-        20: "psk-sha384",
-        21: "pasn",
-        22: "ft-802.1x-sha384",
-        23: "802.1x-sha384",
-        24: "sae-pmk384",
-        25: "ft-sae-pmk384",
-        26: "pasn-defined-key-wrap",
-        29: "edpke",
-    }.get(selector[3])
+# The RSN suite-selector helper cluster (``_rsn_suite_from_layer`` /
+# ``_rsn_suite_list_from_layer`` / ``_rsn_suite_from_oui_type`` /
+# ``_rsn_suite_from_value`` / ``_rsn_suite_selector`` / ``_rsn_cipher_label`` /
+# ``_rsn_akm_label``) moved to ``protocols/wifi.py`` with the ``_normalize_rsn``
+# tshark normalizer it serves.
 
 
 def _field_key(fields: dict[str, JSONObject], layer_name: str) -> str:
