+++ tools/oracle/tests/test_protocol_plugin_coverage.py
@@ -43,8 +43,6 @@ KNOWN_UNMIGRATED: frozenset[str] = frozenset(
         "ble_adv",
         "ble_radio",
         "dot11",
-        "eapol",
-        "rsn",
     }
 )
